Defined common Controller interfaces across server and client
